Depend less on Mod-Tap
Mod-Tap turned out to be unreliable because I'm used to pressing
modifier keys (e.g. CMD) really briefly and so combos like CMD+Q just
wouldn't get registered most of the time.

These changes also include a Meh key, which will be used to invoke
LaunchBar (Meh+Space).

The Backspace key was moved one row lower so it is easier to hit.

The coding layer was reduced to 4 simple keys so they are easier to
remember.

As an experiment, the coding layer can also be enabled by holding the
semicolon key.